aggregate cems based on capacity fuels
